refactor: Extracts styles into separate files named after each component
