+++ client/src/components/Modals/ViewToDoItemModal.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import BaseModal from './BaseModal';
-import { ItemDetails, DetailRow, Label, Value, Divider } from '../../shared/StyledComponents';
+import { ItemDetails, DetailRow, Value, Divider } from '../../styles/ViewToDoItemModal.styles';
+import { Label } from '../../styles/sharedStyles';
 
 const ViewToDoItemModal = ({ isOpen, onRequestClose, item }) => (
   <BaseModal isOpen={isOpen} onRequestClose={onRequestClose} title={item.title}>
