+++ client/src/components/Sidebar.js
@@ -1,33 +1,7 @@
 import React from 'react';
-import styled from 'styled-components';
 import { FiPlus, FiList } from 'react-icons/fi';
+import {SidebarContainer, NavItem } from '../styles/Sidebar.styles';
 
-const SidebarContainer = styled.div`
-  width: 250px;
-  background-color: #2a2a2a;
-  color: #fff;
-  min-height: 100vh;
-`;
-
-const NavItem = styled.div`
-  padding: 15px 20px;
-  cursor: pointer;
-  transition: background-color 0.2s;
-  color: ${(props) => (props.active ? '#000' : '#fff')};
-  background-color: ${(props) => (props.active ? '#90a043' : 'transparent')};
-
-  &:hover {
-    background-color: #6d6d6d;
-    color: #fff;
-  }
-
-  display: flex;
-  align-items: center;
-
-  svg {
-    margin-right: 10px;
-  }
-`;
 
 const Sidebar = ({ activeItem, onItemSelect }) => {
   return (
