+++ client/src/styles/App.styles.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+export const AppContainer = styled.div`
+  display: flex;
+`;
+
+export const ContentArea = styled.div`
+  flex-grow: 1;
+  padding: 20px;
+  display: flex;
+  justify-content: center;
+`;
+
+export const ContentWrapper = styled.div`
+  width: 90%;
+`;
