+++ client/src/styles/EditToDoItemModal.styles.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+export const Select = styled.select`
+  padding: 8px;
+  margin-top: 5px;
+  margin-bottom: 15px;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  width: 100%;
+  box-sizing: border-box;
+`;
+
+export const ButtonContainer = styled.div`
+  display: flex;
+  justify-content: flex-end;
+`;
